% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{aov_t}
\alias{aov_t}
\title{Checking statistically significantly mean difference across the groups}
\usage{
aov_t(data, x, y)
}
\arguments{
\item{data}{a dataframe}

\item{x}{the continues variable to test weather its average is significantly difference across the groups}

\item{y}{the categorical variable to group}
}
\value{
a summary of anova test
}
\description{
To check weather there is statistically significantly difference across the groups.
}
\examples{
aov_t(sleepdurationDay,"SleepDuration", "DayType")
}
