% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicstats.R
\name{basic_stats}
\alias{basic_stats}
\title{Basic Statistics by Group}
\usage{
basic_stats(data, x, y, f)
}
\arguments{
\item{data}{a dataframe}

\item{x}{the variable name you want to caculate on}

\item{y}{the variable name group by, which shoule be a categorical variable}

\item{f}{the basic r function you want to use to perform on the x variable}
}
\description{
Basic Statistics by Group
}
\examples{
## Continues variable stats
basic_stats(sleepdurationDay, "Weight", "DayType", max)
basic_stats(sleepdurationDay, "Weight", "DayType", mean)
basic_stats(sleepdurationDay, "Weight", "DayType", median)

## Categorical variable stats
basic_stats(sleepdurationDay, "Weather", "DayType", table)

}
