% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_of_day.R
\name{classifyday}
\alias{classifyday}
\title{Classify the Day Type}
\usage{
classifyday(df, x, v1 = "weekday", v2 = "weekend")
}
\arguments{
\item{df}{a dataframe}

\item{x}{The variable want to classify}

\item{v1}{The tag name for weekday}

\item{v2}{The tag name for weekend}
}
\value{
a vector
}
\description{
Classfify the variable by type "weekday" and  "weekend"
}
\examples{
classifyday(sleepduration,"DayofWeek")

}
