% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paircorrelation.R
\name{cor_plot}
\alias{cor_plot}
\title{Correlation Matrix Visualization}
\usage{
cor_plot(data, digits = 2)
}
\arguments{
\item{data}{a dataframe}

\item{digits}{the correlation value round to the given digits, by default the digits equal to 2}
}
\value{
a plot
}
\description{
To calculate and visualize all the correlation between two continues variable in a given dataset.
}
\examples{
paircorr(sleepduration)
}
