% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_modelplot.R
\name{log_modelplot}
\alias{log_modelplot}
\title{Simple Linear Regression Model Plot(Log transformed)}
\usage{
log_modelplot(data, x, y)
}
\arguments{
\item{data}{a dataframe}

\item{x}{the independent variable going on the x axis}

\item{y}{the dependent variable going on the y axis}
}
\value{
a plot
}
\description{
Simple Linear Regression Model Plot(Log transformed)
}
\examples{
log_modelplot(sleepdata,"Tempreture", "SleepDuration")

}
\seealso{
\code{\link{modelplot}}
}
