% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelplot.R
\name{modelplot}
\alias{modelplot}
\title{Simple Linear Regression Model Plot}
\usage{
modelplot(data, x, y)
}
\arguments{
\item{data}{a dataframe}

\item{x}{the independent variable going on the x axis}

\item{y}{the dependent variable going on the y axis}
}
\value{
a plot
}
\description{
Simple Linear Regression Model Plot
}
\examples{
modelplot(sleepdata,"Tempreture", "SleepDuration")

}
\seealso{
\code{\link{log_modelplot}}
}
