% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_By_Group.R
\name{plot_box}
\alias{plot_box}
\title{Box Plot by Group}
\usage{
plot_box(data, x, y)
}
\arguments{
\item{data}{a dataframe}

\item{x}{The vairbale on the x axis and the box plot group by this variable}

\item{y}{The variale on the y axis}
}
\value{
a plot
}
\description{
Box Plot by Group
}
\examples{
plot_box(sleepdurationDay, x = "DayType", y= "SleepDuration")

}
