% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatterplot.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter Plot by Group}
\usage{
plot_scatter(data, x, y, color)
}
\arguments{
\item{data}{a dataframe}

\item{x}{The variable is going on the x axis}

\item{y}{The variable is going on the y axis}

\item{color}{color by this variable}
}
\value{
a plot
}
\description{
Scatter Plot by Group
}
\examples{
plot_scatter(sleepdata,"Weight","SleepDuration","DayType")
}
