% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepdata.R
\docType{data}
\name{sleepdata}
\alias{sleepdata}
\title{3 month of sleep and health records of one person}
\format{A data frame with 98 rows and 8 variables}
\source{
gooogle sheet \url{https://docs.google.com/spreadsheets/d/1KdOw0FIxY2Kt8UgY16F3UMQSC7KJaBgtmewNZUGuLY8/edit#gid=0}
weather underground \url{https://www.wunderground.com/history/airport/KROC/2013/12/17/MonthlyHistory.html}
}
\description{
A dataset containing the sleep duration and other attributes during the period from Sep.11 2017 to Dec.17 2017
  The variables are as follows:
}
\details{
\itemize{
  \item Date. in format of month/date/year started from Sep.11 2017
  \item Weight. body weight in kg measured at erery morning when wake up using the same weighing machine
  \item SleepDuration. total sleep time (hour) per night measured by "Sleep Cycle"" app on iphone
  \item Tempreture. daily average tempreture in Celsius from website weather underground
  \item ActivitybySteps. a measurement of activity in steps via iphone 'health' app
  \item Heart Rate.  measured every morning when wake up via "Sleep Cycle"" app on iphone as in beats per minute (bpm)
  \item HavingDog. weather you have dog or not
  \item DayType. he type of the day tag in weekday or weekend
}
}
\keyword{datasets}
