% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepdurationDay.R
\docType{data}
\name{sleepdurationDay}
\alias{sleepdurationDay}
\title{3 week of sleep and health records with day type}
\format{A data frame with 21 rows and 9 variables}
\source{
\url{https://docs.google.com/spreadsheets/d/1KdOw0FIxY2Kt8UgY16F3UMQSC7KJaBgtmewNZUGuLY8/edit#gid=0}
}
\description{
A dataset containing the sleep duration and other attributes during the time of 3 week
  The variables are as follows:
}
\details{
\itemize{
  \item Date. date in format of month/date/year from Sep.11 2017 to Oct.01 2017
  \item DayofWeek. the name of the day of the week (Mon,Tue,Wed,Thu,Fri,Sat,Sun)
  \item Weight. body weight in kg (43.5kg -- 45.9 kg)
  \item SleepDuration.  total sleep time per night in hours(5.33 hour -- 12.00 hour)
  \item Weather. the condition of the weather(Sunny, Cloudy, Rainy, Snowy)
  \item Tempreture.  tempreture in Celsius (12.00--25.00)
  \item ActivitybySteps. a measurement of activity in steps (16--8331)
  \item Heart Rate. heart rate in beats per minute (66 bpm--93 bpm)
  \item DayType. the type of the day tag in weekday or weekend
}
}
\keyword{datasets}
