% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slm_log.R
\name{slm_log}
\alias{slm_log}
\title{Fit a Simple Linear Regression Model (Log transformed)}
\usage{
slm_log(data, x, y)
}
\arguments{
\item{data}{a dataframe}

\item{x}{the independent variable}

\item{y}{the dependent varible}
}
\value{
summary of the model
}
\description{
This is a function to conduct simple linear regression model on two continues variable,
Since the dependent variable may violate the normality assumption, so in this function, the dependent
variable is log transformed to let it follwo normal distrbution.
}
\examples{
slm_log(sleepduration, "Tempreture", "SleepDuration")


}
\seealso{
\code{\link{slm_s}}
}
\author{
Yunrou Gong
}
