% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slm_s.R
\name{slm_s}
\alias{slm_s}
\title{Fit a Simple Linear Regression Model}
\usage{
slm_s(data, x, y)
}
\arguments{
\item{data}{a dataframe}

\item{x}{the independent variable on x axis}

\item{y}{the dependent variable on y axis}
}
\value{
summary of the model
}
\description{
This is a function to conduct simple linear regression model on two continues variable
}
\examples{
slm_s(sleepduration, "Tempreture", "Heart Rate")
}
\seealso{
\code{\link{slm_log}}
}
\author{
Yunrou Gong
}
